# contracts/api-openapi.yaml
openapi: 3.0.3
info:
  title: South African RailWaze API
  description: Core backend service contracts for Iteration 2
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local Development Server
paths:
  /waypoints:
    get:
      summary: Retrieve all anchor waypoints
      description: Returns the list of 4 key anchor waypoints along the rail route.
      operationId: getWaypoints
      responses:
        '200':
          description: Successful retrieval of waypoints
          content:
            application/json:
              schema:
                $ref: './waypoint.schema.json'

  /waypoints/{id}/trivia:
    get:
      summary: Get trivia questions for a waypoint
      description: Fetches localized cultural and historical trivia for a given waypoint.
      operationId: getWaypointTrivia
      parameters:
        - name: id
          in: path
          required: true
          description: Waypoint identifier
          schema:
            type: string
      responses:
        '200':
          description: Successful trivia payload
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './trivia-passport.schema.json#/definitions/TriviaQuestionPublic'
        '404':
          description: Waypoint not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string

  /waypoints/{waypoint_id}/trivia/answer:
    post:
      summary: Submit a trivia answer
      description: Validates the original option index server-side and awards a station stamp for a correct answer.
      operationId: submitWaypointTriviaAnswer
      parameters:
        - name: waypoint_id
          in: path
          required: true
          description: Waypoint identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './trivia-passport.schema.json#/definitions/AnswerSubmission'
      responses:
        '200':
          description: Answer result and optional stamp award
          content:
            application/json:
              schema:
                $ref: './trivia-passport.schema.json#/definitions/AnswerResult'
        '404':
          description: Waypoint or trivia question not found
        '429':
          description: Too many answer submissions

  /passport/{sessionId}:
    get:
      summary: Get user passport state
      description: Fetches active user rank, collected stamps, and completed trivia.
      operationId: getPassportBySession
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Unique traveler session identifier (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current user passport and journey progression
          content:
            application/json:
              schema:
                $ref: './trivia-passport.schema.json#/definitions/PassportState'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
    post:
      summary: Rejected - rank and stamps cannot be set by the client
      description: >-
        STRIDE "Elevation of Privilege" mitigation (issue #18). Traveler
        rank and stamps are computed exclusively server-side from
        verified trivia answers (POST /waypoints/{waypointId}/trivia/answer).
        This method is intentionally not supported on this path and always
        returns 405, regardless of the request body.
      operationId: rejectPassportMutation
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Unique traveler session identifier (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '405':
          description: Rank/stamps cannot be set directly by the client
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string